+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 ## Unreleased
 
+### Fixed
+- `clk_rst_gen`: Fix generation of odd clock periods.
+
 ## v0.2.0 - 2019-08-20
 
 ### Added