+++ CHANGELOG.md
@@ -4,12 +4,14 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Unreleased
+
+## 0.2.1 - 2021-12-02
 
 ### Fixed
 - `clk_rst_gen`:
   - Fix generation of odd clock periods.
   - Fix number of reset cycles.
+
 
 ## v0.2.0 - 2019-08-20
 
@@ -22,17 +24,20 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
   the declarations in favor of a simulation-wide precision declaration.
 - Rename parameters to comply with style guidelines.
 
+
 ## v0.1.2 - 2019-08-20
 
 ### Fixed
 - rand_synch_driver: Fix instantiation of `rand_synch_holdable_driver`.
 - rand_stream_slv: Fix instantiation of `rand_sync_driver`.
 
+
 ## v0.1.1 - 2019-02-26
 
 ### Fixed
 - Move all files into the `simulation` target. This precludes synthesis of files in this package
   when this package is included as dependency.
+
 
 ## v0.1.0 - 2019-02-25